+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Migrated `rector.php` to the `RectorConfig::configure()` builder.
 - Gathered all repositories in one Symfony application.
 - Changed to vite 7 and rolldown.
 - Added ADRs 008 and 009.